Adding get-foldersize and get-userhomefolderinfo
get-foldersize recursively searched the child folders and files and adds up the size in bytes

get-userhomefolderinfo takes in a root folder, and for every user folder checks to see if there is a corresponding ad user. if so, add up the folder size of their home folder. If no user exists in AD, output the folder as an orphan.